NMEAParser

Documentation for NMEAParser.

NMEAParser.GGA

NMEAParser.NMEAData

NMEAParser.TWPOS

NMEAParser._dms_to_dd

NMEAParser._hms_to_secs

NMEAParser.get_system

NMEAParser.parse

NMEAParser.parse_GGA

NMEAParser.parse_TWPOS

NMEAParser.parse_msg!

NMEAParser.pos_convert

NMEAParser.vel_convert

NMEAParser.GGA — Type

GGA(sys::AbstractString)

A mutable struct representing a GGA (Global Positioning System Fix Data) message.

Fields

system: GPS, GLONASS, GALILEO, or Combined.

time: Time in seconds.

latitude: Latitude in decimal degrees.

longitude: Longitude in decimal degrees.

fix_quality: Quality of the fix.

num_sats: Number of satellites used in the fix.

HDOP: Horizontal Dilution of Precision.

altitude: Altitude above Mean Sea Level in meters.

geoidal_seperation: Geoidal separation in meters.

age_of_differential: Age of differential data in seconds since last SC104 type 1 or 9 update.

diff_reference_id: Differential reference station ID.

valid: Whether the data is valid.

Constructor

The constructor takes one argument:

sys::AbstractString: The system type (GPS, GLONASS, GALILEO, or Combined).

Examples

gga = GGA("GPS")

NMEAParser.NMEAData — Type

NMEAData()

A mutable struct representing NMEA data.

Fields

last_GGA::GGA: The last GGA message received.

last_RMC::RMC: The last RMC message received.

last_GSA::GSA: The last GSA message received.

last_GSV::GSV: The last GSV message received.

last_GBS::GBS: The last GBS message received.

last_VTG::VTG: The last VTG message received.

last_GLL::GLL: The last GLL message received.

last_ZDA::ZDA: The last ZDA message received.

last_DTM::DTM: The last DTM message received.

last_PASHR::PASHR: The last PASHR message received.

last_TWPOS::TWPOS: The last TWPOS message received.

Constructor

The constructor takes no arguments and initializes all fields to their respective types with the system set to "UNKNOWN".

Examples

nmea_data = NMEAData()

NMEAParser.TWPOS — Type

mutable struct TWPOS

A mutable structure representing a TWPOS sentence from a GPS system.

Fields

system: the GPS system

time: the time of the sentence

xpose: the x position

ypose: the y position

zpose: the z position

distance: the distance

velocity: the velocity

valid: a boolean indicating if the sentence is valid

Constructor

TWPOS(sys::AbstractString)

Constructs a new instance of the TWPOS type with the given GPS system.

Example

sys = "GPS"
TWPOS_data = TWPOS(sys)

NMEAParser._dms_to_dd — Method

_dms_to_dd(dms::SubString, hemi::SubString)

Converts a string representing degrees, minutes and seconds (DMS) to decimal degrees.

Arguments

dms: a substring representing degrees, minutes and seconds

hemi: a substring representing the hemisphere

Returns

dec_degrees: the decimal degree representation of the input DMS

Example

dms = "4807.038"
hemi = "N"
dec_degrees = _dms_to_dd(dms, hemi)

NMEAParser._hms_to_secs — Method

_hms_to_secs(hms::SubString)

Converts a string representing hours, minutes and seconds (HMS) to seconds.

Arguments

hms: a substring representing hours, minutes and seconds

Returns

seconds: the number of seconds represented by the input HMS

Example

hms = "123519"
seconds = _hms_to_secs(hms)

NMEAParser.get_system — Method

get_system(mtype::SubString)

Determine the system type based on the input mtype.

Arguments

mtype: A SubString representing the message type.

Returns

A string representing the system type. Possible values are "GPS", "GLONASS", "GALILEO", "COMBINED", and "UNKNOWN".

Examples

get_system("$GPGGA") # "GPS"
get_system("$GLGSV") # "GLONASS"
get_system("$GAGGA") # "GALILEO"
get_system("$GNGNS") # "COMBINED"
get_system("$PMTK")  # "UNKNOWN"

NMEAParser.parse — Method

parse(line::AbstractString)

Parses an NMEA sentence, returning a corresponding type.

Arguments

line::AbstractString

NMEAParser.parse_GGA — Method

parse_GGA(items::Array{T}, system::AbstractString) where T <: SubString

Parse an array of substrings items representing a GGA sentence from a GPS system.

Arguments

items: an array of substrings representing the fields of a GGA sentence

system: a string representing the GPS system

Returns

GGA_data: an instance of the GGA type containing the parsed data

Example

items = ["$GPGGA", "123519", "4807.038", "N", "01131.000", "E", "1", "08", "0.9", "545.4", "M", "46.9", "M", "", ""]
system = "GPS"
GGA_data = parse_GGA(items, system)

NMEAParser.parse_TWPOS — Method

parse_TWPOS(items::Array{T}, system::AbstractString) where T<:SubString

Parse an array of substrings items representing a TWPOS sentence from a GPS system.

Arguments

items: an array of substrings representing the fields of a TWPOS sentence

system: a string representing the GPS system

Returns

TWPOS_data: an instance of the TWPOS type containing the parsed data

Example

items = ["$GPTWPOS", "123519", "4807.038", "M", "01131.000", "M", "545.4", "M", "46.9", "M", "0.9", "K"]
system = "GPS"
TWPOS_data = parse_TWPOS(items, system)

NMEAParser.parse_msg! — Method

parse_msg!(s::NMEAData, line::AbstractString)

Parses an NMEA sentence and updates the corresponding field in the NMEAData struct.

Arguments

s::NMEAData: The NMEAData struct to update.

line::AbstractString: The NMEA sentence to parse.

Returns

Returns the type of message parsed as a string.

Examples

nmea_data = NMEAData()
parse_msg!(nmea_data, "$GPGGA,123519,4807.038,N,01131.000,E,1,08,0.9,545.4,M,46.9,M,,*47")

Notes

The function supports parsing of DTM, GBS, GGA, GLL, GNS, GSA, GSV, RMC, VTG, ZDA, PASHR and TWPOS messages.

NMEAParser.pos_convert — Method

pos_convert(Flag, Value)

Convert a position Value to a different measurement system based on flag.

NMEAParser.vel_convert — Method

vel_convert(Flag, Value)

Convert a velocity Value to a different measurement system based on flag.